Remove old dnscheck implementation. Added TODOs
