+++ whois-commons/src/main/resources/whois.properties
@@ -90,10 +90,6 @@ mailupdates.database.url=jdbc:mariadb://localhost/MAILUPDATES_LOCAL
 mailupdates.database.username=dbint
 mailupdates.database.password=
 
-dnscheck.database.url=jdbc:mariadb://localhost/DNSCHECK_LOCAL
-dnscheck.database.username=dbint
-dnscheck.database.password=
-
 acl.database.url=jdbc:mariadb://localhost/ACL_LOCAL
 acl.database.username=dbint
 acl.database.password=
