+++ whois-endtoend/src/test/resources/whois.properties
@@ -89,10 +89,6 @@ mailupdates.database.url=jdbc:mysql://localhost/MAILUPDATES_LOCAL
 mailupdates.database.username=dbint
 mailupdates.database.password=
 
-dnscheck.database.url=jdbc:mysql://localhost/DNSCHECK_LOCAL
-dnscheck.database.username=dbint
-dnscheck.database.password=
-
 acl.database.url=jdbc:mysql://localhost/ACL_LOCAL
 acl.database.username=dbint
 acl.database.password=
