+++ whois-update/src/main/java/net/ripe/db/whois/update/dns/zonemaster/ZonemasterDnsGateway.java
@@ -30,6 +30,7 @@ import java.util.stream.Collectors;
 
 import static net.ripe.db.whois.common.Messages.Type.ERROR;
 
+// TODO: [ES] handle failure cases (e.g. server down, timeouts etc).
 // TODO: [ES] DnsCheckRequest glue is never read, is it needed?
 @DeployedProfile
 //@Primary      // TODO: [ES] DnsGatewayImpl is deployed instead
