+++ whois-update/src/test/java/net/ripe/db/whois/update/dns/zonemaster/ZonemasterDnsGatewayTest.java
@@ -7,6 +7,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
+// TODO: [ES] implement tests
 @RunWith(MockitoJUnitRunner.class)
 public class ZonemasterDnsGatewayTest {
